pyproject.toml
[project]
name = "TheWorld"
version = "0.1.0"
description = "Fused vision-language-world model combining Gemma 3 and Cosmos"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "diffusers>=0.34.0",
    "peft>=0.15.0",
    "torch",
    "torchaudio",
    "torchvision",
    "transformers>=4.56.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "evaluate>=0.4.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "huggingface-hub>=0.20.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "wandb>=0.18.0",
    "tensorboard>=2.18.0",
]

[tool.black]
line-length = 120
target-version = ["py312"]
